% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R2}
\alias{R2}
\alias{R2adj}
\alias{R2adj2}
\alias{R2fpe}
\alias{R2gcv}
\alias{R2press}
\title{R squared}
\usage{
R2(object, ...)

R2adj(object, ...)

R2adj2(object, ...)

R2gcv(object, ...)

R2press(object, ...)

R2fpe(object, ...)
}
\arguments{
\item{object}{Fitted model object}

\item{...}{Additional parameters}
}
\description{
Calculate R squared statistics from fitted model objects
}
\details{
So far there are six different types of R squared statistics.
}

