% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquared.R
\name{Rsquared}
\alias{Rsquared}
\title{Rsquared Function}
\usage{
Rsquared(x, y, xname, title)
}
\arguments{
\item{x}{x = time, which can be any unit, but must be numeric}

\item{y}{y = absorbance values in a dataset; must be numeric}

\item{xname}{xname allows users to change the x-axis label; can be character and/or numeric}

\item{title}{title allows users to change the graph title; can be character and/or numeric}
}
\value{
graph of time versus absorbance means on a log scale and the slope of the graph, the R-squared value, and other coefficient values.
}
\description{
The Rsquared function calculates the mean of replicate absorbances, graphs time versus the absorbance mean in log scale, and returns to users the slope of the graph, the R-squared value, and other coefficient values.
}
\keyword{absorbance}
